+++ ibig/src/bytes.rs
@@ -43,6 +43,7 @@ impl UBig {
     /// assert_eq!(UBig::from(0x0105u16).to_be_bytes(), [1, 5]);
     /// assert_eq!(UBig::from(0u8).to_be_bytes(), []);
     /// ```
+    #[inline]
     pub fn to_be_bytes(&self) -> Vec<u8> {
         // Big-endian is the little-endian representation reversed.
         let mut bytes = self.to_le_bytes();
@@ -59,6 +60,7 @@ impl UBig {
     /// # use ibig::UBig;
     /// assert_eq!(UBig::from_le_bytes(&[5, 1]), UBig::from(0x0105u16));
     /// ```
+    #[inline]
     pub fn from_le_bytes(bytes: &[u8]) -> UBig {
         let mut digits = Digits::new();
         digits.resize(bytes.len().div_ceil(Digit::BYTES), Digit::ZERO);
@@ -90,6 +92,7 @@ impl UBig {
     /// # use ibig::UBig;
     /// assert_eq!(UBig::from_be_bytes(&[1, 5]), UBig::from(0x0105u16));
     /// ```
+    #[inline]
     pub fn from_be_bytes(bytes: &[u8]) -> UBig {
         let mut digits = Digits::new();
         digits.resize(bytes.len().div_ceil(Digit::BYTES), Digit::ZERO);
@@ -138,6 +141,7 @@ impl IBig {
     /// assert_eq!(IBig::from(0xffffi32).to_be_bytes(), [0, 0xff, 0xff]);
     /// assert_eq!(IBig::from(-1i8).to_be_bytes(), [0xff]);
     /// ```
+    #[inline]
     pub fn to_be_bytes(&self) -> Vec<u8> {
         let mut bytes = self.to_le_bytes();
         bytes.reverse();
@@ -158,6 +162,7 @@ impl IBig {
     /// assert_eq!(IBig::from_le_bytes(&[1, 2, 0]), IBig::from(0x0201i16));
     /// assert_eq!(IBig::from_le_bytes(&[0xff]), IBig::from(-1i8));
     /// ```
+    #[inline]
     pub fn from_le_bytes(bytes: &[u8]) -> IBig {
         let mut digits = Digits::new();
         digits.resize(bytes.len().div_ceil(Digit::BYTES), Digit::ZERO);
@@ -195,6 +200,7 @@ impl IBig {
     /// assert_eq!(IBig::from_be_bytes(&[0, 1, 2]), IBig::from(0x0102i16));
     /// assert_eq!(IBig::from_be_bytes(&[0xff]), IBig::from(-1i8));
     /// ```
+    #[inline]
     pub fn from_be_bytes(bytes: &[u8]) -> IBig {
         let mut digits = Digits::new();
         digits.resize(bytes.len().div_ceil(Digit::BYTES), Digit::ZERO);